requirements.txt
gspread==5.9.0
google-auth==2.22.0
requests-oauthlib==1.3.1
python-dateutil==2.8.2
pandas==2.2.2
